feat(app): add Layout for pages
