+++ src/app/(main)/about/page.tsx
@@ -1,9 +1,8 @@
 import { Header } from "@/widgets/header";
 
-export default function OfferAgreement() {
+export default function About() {
   return (
     <div>
-      <Header />
       <main></main>
     </div>
   );
