+++ src/app/(main)/blog/page.tsx
@@ -1,9 +1,8 @@
 import { Header } from "@/widgets/header";
 
-export default function About() {
+export default function Blog() {
   return (
     <div>
-      <Header />
       <main></main>
     </div>
   );
