+++ src/app/(main)/gallery/page.tsx
@@ -1,9 +1,8 @@
 import { Header } from "@/widgets/header";
 
-export default function Reviews() {
+export default function Gallery() {
   return (
     <div>
-      <Header />
       <main></main>
     </div>
   );
