+++ src/app/(main)/layout.tsx
@@ -1,0 +1,14 @@
+import { Footer } from "@/widgets/footer";
+import { Header } from "@/widgets/header";
+
+type LayoutProps = { children: React.ReactNode };
+
+export default function Layout({ children }: LayoutProps) {
+  return (
+    <div>
+      <Header />
+      {children}
+      <Footer />
+    </div>
+  );
+}
