+++ src/app/(main)/location/page.tsx
@@ -1,9 +1,8 @@
 import { Header } from "@/widgets/header";
 
-export default function Policy() {
+export default function Locations() {
   return (
     <div>
-      <Header />
       <main></main>
     </div>
   );
