+++ src/app/(main)/page.tsx
@@ -2,14 +2,11 @@ import { Advantages } from "@/widgets/advantages";
 import { Banner } from "@/widgets/banner";
 import { Hero } from "@/widgets/hero";
 import { Faq } from "@/widgets/faq";
-import { Footer } from "@/widgets/footer";
-import { Header } from "@/widgets/header";
 import { ContactForm } from "@/widgets/contact-form";
 
 export default function Home() {
   return (
     <div>
-      <Header />
       <main>
         <Hero />
         <Advantages />
@@ -17,7 +14,6 @@ export default function Home() {
         <Faq />
         <Banner />
       </main>
-      <Footer />
     </div>
   );
 }
