+++ src/app/(main)/policy/page.tsx
@@ -1,9 +1,8 @@
 import { Header } from "@/widgets/header";
 
-export default function Blog() {
+export default function Policy() {
   return (
     <div>
-      <Header />
       <main></main>
     </div>
   );
