+++ src/app/(main)/reviews/page.tsx
@@ -1,9 +1,8 @@
 import { Header } from "@/widgets/header";
 
-export default function Locations() {
+export default function Reviews() {
   return (
     <div>
-      <Header />
       <main></main>
     </div>
   );
